api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.133.0
h11==0.16.0
idna==3.11
passlib==1.7.4
joblib==1.5.3
neo4j==6.1.0
numpy==2.4.2
pandas==3.0.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
regex==2026.2.19
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.3
urllib3==2.6.3
uvicorn==0.41.0

